+++ cavalier_contours/src/core/math/base_math.rs
@@ -260,28 +260,31 @@ where
 /// Returns the parametric value on the line segment going from `p0` to `p1` at the `point` given.
 ///
 /// Note this function assumes the `point` is on the line and properly handles the cases of vertical
-/// and horizontal lines by using the `epsilon` parameter to fuzzy compare for when `p0.x == p1.x`.
+/// and horizontal lines by using the component with the largest difference in the calculation.
 #[inline]
 pub fn parametric_from_point<T>(p0: Vector2<T>, p1: Vector2<T>, point: Vector2<T>, epsilon: T) -> T
 where
     T: Real,
 {
-    if p0.x.fuzzy_eq_eps(p1.x, epsilon) {
-        // vertical segment, use y coordinate
-        debug_assert!(
-            point.x.fuzzy_eq_eps(p0.x, epsilon),
-            "point does not lie on the line defined by p0 to p1"
-        );
-        (point.y - p0.y) / (p1.y - p0.y)
+    let x_diff = p1.x - p0.x;
+    let y_diff = p1.y - p0.y;
+
+    debug_assert!(
+        (((p1.x - p0.x) * (p0.y - point.y) - (p0.x - point.x) * (p1.y - p0.y)).abs()
+            / (x_diff * x_diff + y_diff * y_diff).sqrt())
+        .fuzzy_eq_zero_eps(epsilon),
+        "point does not lie on the line defined by p0 to p1 (based on distance)"
+    );
+
+    // use larger diff component for calculation to avoid problems with vertical/horizontal lines
+    // where component diff is zero and to avoid adding error in calculation caused by small
+    // denominator
+    if x_diff.abs() < y_diff.abs() {
+        // use y component
+        (point.y - p0.y) / y_diff
     } else {
-        // use x coordinate
-        debug_assert!(
-            point.fuzzy_eq_eps(p0, epsilon)
-                || ((point.y - p0.y) / (point.x - p0.x))
-                    .fuzzy_eq_eps((p1.y - p0.y) / (p1.x - p0.x), epsilon),
-            "point does not lie on the line defined by p0 to p1"
-        );
-        (point.x - p0.x) / (p1.x - p0.x)
+        // use x component
+        (point.x - p0.x) / x_diff
     }
 }